requirements.txt
streamlit
pandas
requests
openai
python-dotenv
plotly